analysis/requirements.txt
pandas>=1.5
matplotlib>=3.6
statsmodels>=0.14
python-pptx>=0.6.21
